+++ embedded/android/PhoneController/app/src/main/java/com/ndomx/phonecontroller/api/Command.kt
@@ -1,4 +1,4 @@
-package com.ndomx.phonecontroller.contracts
+package com.ndomx.phonecontroller.api
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.JsonElement
