+++ embedded/android/PhoneController/app/src/main/java/com/ndomx/phonecontroller/api/ExecuteCommandResult.kt
@@ -1,4 +1,4 @@
-package com.ndomx.phonecontroller.contracts
+package com.ndomx.phonecontroller.api
 
 enum class ExecuteCommandResult(val value: Int) {
     OK(0),
